notification-service/requirements.txt
aerich==0.7.2
aiokafka==0.9.0
aiosqlite==0.17.0
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
dictdiffer==0.9.0
fastadmin==0.1.41
fastapi==0.68.0
greenlet==3.0.2
h11==0.14.0
iso8601==1.1.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
packaging==23.2
parso==0.8.3
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptpython==3.0.25
py-postgresql==1.3.0
pydantic==1.10.8
Pygments==2.17.2
PyJWT==2.8.0
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3.post1
starlette==0.14.2
tomlkit==0.12.3
tortoise-cli==0.1.2
tortoise-orm==0.20.0
typing_extensions==4.9.0
uvicorn==0.15.0
wcwidth==0.2.12
